Restructure interfaces into clearer abstraction folders
