+++ src/EduConnect.Api/Abstractions/ServicesAbstractions/IStudentsService.cs
@@ -1,6 +1,6 @@
 using EduConnect.Api.Entities;
 
-namespace EduConnect.Api.Abstractions;
+namespace EduConnect.Api.Abstractions.ServicesAbstractions;
 
 public interface IStudentsService
 {
