+++ src/EduConnect.Api/Repositories/TeacherRepository.cs
@@ -1,4 +1,4 @@
-using EduConnect.Api.Abstractions;
+using EduConnect.Api.Abstractions.RepositoriesAbstractions;
 using EduConnect.Api.Data;
 using EduConnect.Api.Entities;
 using EduConnect.Api.Exceptions;
