+++ src/EduConnect.Api/Services/ClassesService.cs
@@ -1,7 +1,7 @@
-using EduConnect.Api.Abstractions;
+using EduConnect.Api.Abstractions.RepositoriesAbstractions;
+using EduConnect.Api.Abstractions.ServicesAbstractions;
 using EduConnect.Api.Entities;
 using EduConnect.Api.Exceptions;
-using Microsoft.Extensions.Logging;
 
 namespace EduConnect.Api.Services;
